Add delay for CharacterCount story in Happo (#950)
I'm currently debugging a spurious issue with emojis sometimes not
showing up in the ios-safari target on happo.io. This is an example
of such a spurious diff:
https://happo.io/a/371/p/453/compare/e934ff76a58cdbc523eef0a3c496fe2d20c02c73/cf58cd0cc54d7da04ce184d380b6f73da64a439e

My attempts at isolating this issue and reproduce haven't played out.
I've tried rendering roughly the same html as in the CharacterCount
story but all my screenshots come back with visible emoji characters.
This makes me wonder if there is maybe a race condition where Safari on
iOS is doing some optimization where a first render isn't showing emoji
(but the second/nth render does). If that's the case, adding a short
delay before Happo takes the screenshot should help resolve this.

Co-authored-by: Henric Trotzig <henric.trotzig@gmail.com>