Split up Internet Explorer runs in 3 for Happo tests
We've recently seen a number of timeouts caused by Internet Explorer
taking too long to process jobs.

Internet Explorer is going away in less than 2 months, and over at Happo
we've found it increasingly hard to get it to perform well, especially
for Storybook test suites.

The most pragmatic approach might be to remove it completely from
.happo.js. As a temporary workaround however, I'm applying the `chunks`
option. It will split up the IE job in three parts, hopefully allowing
it to finish sooner and avoid the timeouts.
https://docs.happo.io/docs/configuration#target-chunks