Move the check into a variable;
I'm doing this since I need to use it in two places, but I think
eventually this should be re-written maybe. For now though, I like the
idea of having a variable named that determines if things should run,
and then check that variable in various places to determine if it needs
to be run or not.