Making some light configuration changes;
These changes were related to errors and warnings I received when running `bun
run lint` and `bun run build` in the project. Specifically, the `bun run lint`
had an issue with `.eslintrc` not having a `"root": true,` line. And the `bun
run build` yielded a bunch of Typescript errors without the `"paths": { "react":
[ "./node_modules/@types/react" ] }` line.