Refactoring the iOS function;
This was throwing all sorts of Typescript errors. I simplified the check a lot
to only use user-agent string checking. It's not as robust, but it removes the
TS errors from the project. Also, `navigator.platform` has been deprecated. From
everything I searched for it seems like when this code was added in early 2021,
the code was acceptable. But now in 2023, the community consensus is to only
check the user-agent and avoid checking for `window.MSStream`.